Reorganize explorer folder structure
