+++ apps/studio/components/interfaces/Explorer/QueryEditor/DisplaySettingsButton.tsx
