+++ apps/studio/components/interfaces/Explorer/QueryEditor/QueryResultError.tsx
@@ -3,8 +3,8 @@ import { ExternalLink } from 'lucide-react'
 import { parseAsBoolean, useQueryState } from 'nuqs'
 import { Button, cn, Tooltip, TooltipContent, TooltipTrigger } from 'ui'
 
-import { subscriptionHasHipaaAddon } from '../Billing/Subscription/Subscription.utils'
-import { type QueryResult } from './types'
+import { subscriptionHasHipaaAddon } from '../../Billing/Subscription/Subscription.utils'
+import { type QueryResult } from '../types'
 import { AiAssistantDropdown } from '@/components/ui/AiAssistantDropdown'
 import CopyButton from '@/components/ui/CopyButton'
 import { InlineLink, InlineLinkClassName } from '@/components/ui/InlineLink'
