+++ apps/studio/components/interfaces/Explorer/__tests__/QueryTab.test.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event'
 import { HttpResponse } from 'msw'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { QueryTab } from './QueryTab'
+import { ExplorerQueryTab } from '../ExplorerQueryTab'
 import type { ReadReplicasData } from '@/data/read-replicas/replicas-query'
 import { explorerQueryState } from '@/state/explorer-query'
 import { createTabsState, TabsStateContext } from '@/state/tabs'
@@ -49,7 +49,7 @@ vi.mock('./ExplorerQuerySourceMenu', () => ({
 const renderQueryTab = () =>
   customRender(
     <TabsStateContext.Provider value={createTabsState('default')}>
-      <QueryTab />
+      <ExplorerQueryTab />
     </TabsStateContext.Provider>
   )
 
@@ -196,7 +196,7 @@ describe('QueryTab execution', () => {
     testContext.params = { ref: 'default', id: 'query-test-2' }
     rerender(
       <TabsStateContext.Provider value={createTabsState('default')}>
-        <QueryTab />
+        <ExplorerQueryTab />
       </TabsStateContext.Provider>
     )
