+++ apps/studio/pages/project/[ref]/explorer/chat/[id].tsx
@@ -1,9 +1,9 @@
-import { ChatEditor } from '@/components/interfaces/Explorer/ChatEditor'
+import { ExplorerChatTab } from '@/components/interfaces/Explorer/ExplorerChatTab'
 import { DefaultLayout } from '@/components/layouts/DefaultLayout'
 import { ExplorerLayout } from '@/components/layouts/ExplorerLayout/ExplorerLayout'
 import type { NextPageWithLayout } from '@/types'
 
-const ChatPage: NextPageWithLayout = () => <ChatEditor />
+const ChatPage: NextPageWithLayout = () => <ExplorerChatTab />
 
 ChatPage.getLayout = (page) => (
   <DefaultLayout>
